refactor(cmd/scan): decompose scan.rs god-file into focused submodules (#1009)
`src/cmd/scan.rs` was 3,830 lines — the largest file in the repo — with
`run_scan` acting as a god-function tangling arg parsing, preflight, target
resolution, the preflight/analysis loop, the scanning loop, PoC rendering,
result post-processing, and output routing into one body.

Split `cmd/scan.rs` into `cmd/scan/` with `run_scan` left as a thin
orchestrator that reads as a sequence of named stages:

- `args.rs`       — `ScanArgs`, default/cap constants, value parsers, `PreflightOptions`
- `validation.rs` — numeric arg checks + URL/file input heuristics
- `preflight.rs`  — content-type / CSP / WAF preflight + reqwest error classification
- `input.rs`      — target resolution (input-type, file/stdin/raw-HTTP, dedup, scope filters, cookies)
- `analysis.rs`   — preflight + parameter-analysis loop
- `scan_loop.rs`  — per-host scanning loop + streaming printer
- `output.rs`     — dry-run / only-discovery / end-of-scan result rendering
- `poc.rs`        — curl/httpie/plain POC + finding-block rendering
- `postprocess.rs`— dedupe / priority / context extraction
- `logging.rs`    — log/spinner helpers (were `run_scan` closures)
- `mod.rs`        — `run_scan` orchestrator + shared `ScanState`

The shared cross-task `Arc` handles move onto a `ScanState` struct; each stage
rebinds them to owned locals up front so the moved-out bodies stay verbatim
from the pre-split `run_scan`. The four `run_scan`-local closures
(`log_info`/`log_warn`/`log_dbg`/`start_spinner`) become free functions in
`logging.rs` taking `&ScanArgs` / `spinner_allowed` explicitly.

No behavior change: the public API (`ScanArgs`, `ScanOutcome`, `run_scan`,
`PreflightOptions`, `DEFAULT_*`/`CLI_MAX_*` constants) is preserved via
re-exports from `mod.rs`. Each extracted stage body byte-diffs against the
original modulo the mechanical changes above.

Acceptance criteria:
- No file in src/cmd/scan/ over ~800 lines (largest non-test: input.rs 640)
- src/cmd/scan/tests.rs (38) + 1173 lib + scan_run_paths (10) +
  functional/integration (189) pass
- cargo clippy --all-targets --all-features -- -D warnings clean

AGENTS.md updated to point at the new module layout.
