+++ src/cmd/scan/tests.rs
@@ -1,9 +1,10 @@
+use super::poc::{build_ast_dom_message, generate_poc};
+use super::postprocess::{dedupe_ast_results, extract_context};
+use super::preflight::{PreflightOutcome, is_allowed_content_type, preflight_content_type};
 use super::{
     CLI_MAX_DELAY_MS, CLI_MAX_TIMEOUT_SECS, CLI_MAX_WORKERS, DEFAULT_DELAY_MS, DEFAULT_ENCODERS,
     DEFAULT_MAX_CONCURRENT_TARGETS, DEFAULT_MAX_TARGETS_PER_HOST, DEFAULT_METHOD,
-    DEFAULT_TIMEOUT_SECS, DEFAULT_WORKERS, PreflightOutcome, ScanArgs, build_ast_dom_message,
-    dedupe_ast_results, extract_context, generate_poc, is_allowed_content_type,
-    preflight_content_type, validate_numeric_args,
+    DEFAULT_TIMEOUT_SECS, DEFAULT_WORKERS, ScanArgs, validate_numeric_args,
 };
 use crate::scanning::result::{FindingType, Result as ScanResult};
 use crate::target_parser::parse_target;
